perf: 채널 맴버Id를 복합키로 변경, Assembler 서비스 코드 및 성능 개선
JoinedChannelAssembler에 캐싱 기능 추가
JoinedChannelResponse 필드를 분리하여 각각의 Assembler에서 진행하는 것으로 구조 개선

+++ src/main/java/com/zonbeozon/channel/dto/ChannelWithMemberCount.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 @Getter
 public class ChannelWithMemberCount {
     private Channel channel;
-    private int memberCount;
+    private Long memberCount;
 }
